transform: split into one module per entry type
