+++ site/zenodo_rdm/migrator/transform/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2023 CERN.
+#
+# ZenodoRDM is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+
+"""Zenodo migrator transformers."""
+
+
+from .communities import ZenodoCommunityTransform
+from .records import ZenodoRecordTransform
+from .users import ZenodoUserTransform
+
+
+__all__ = (ZenodoCommunityTransform, ZenodoRecordTransform, ZenodoUserTransform)
